[DEV-2378] move specialExternalAccount to transaction module
