+++ src/subdomains/core/aml/aml.module.ts
@@ -1,10 +1,11 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { UserModule } from 'src/subdomains/generic/user/user.module';
 import { BankModule } from 'src/subdomains/supporting/bank/bank.module';
+import { TransactionModule } from 'src/subdomains/supporting/payment/transaction.module';
 import { AmlService } from './aml.service';
 
 @Module({
-  imports: [forwardRef(() => UserModule), BankModule],
+  imports: [forwardRef(() => UserModule), TransactionModule, BankModule],
   controllers: [],
   providers: [AmlService],
   exports: [AmlService],
