+++ src/subdomains/supporting/payment/payment.module.ts
@@ -13,9 +13,11 @@ import { TransactionRequest } from './entities/transaction-request.entity';
 import { TransactionSpecification } from './entities/transaction-specification.entity';
 import { BlockchainFeeRepository } from './repositories/blockchain-fee.repository';
 import { FeeRepository } from './repositories/fee.repository';
+import { SpecialExternalAccountRepository } from './repositories/special-external-account.repository';
 import { TransactionRequestRepository } from './repositories/transaction-request.repository';
 import { TransactionSpecificationRepository } from './repositories/transaction-specification.repository';
 import { FeeService } from './services/fee.service';
+import { SpecialExternalAccountService } from './services/special-external-account.service';
 import { TransactionHelper } from './services/transaction-helper';
 import { TransactionRequestService } from './services/transaction-request.service';
 import { TransactionJobModule } from './transaction-job.module';
@@ -42,7 +44,9 @@ import { TransactionModule } from './transaction.module';
     TransactionRequestRepository,
     BlockchainFeeRepository,
     TransactionRequestService,
+    SpecialExternalAccountService,
+    SpecialExternalAccountRepository,
   ],
-  exports: [TransactionHelper, FeeService, TransactionRequestService],
+  exports: [TransactionHelper, FeeService, TransactionRequestService, SpecialExternalAccountService],
 })
 export class PaymentModule {}
