+++ src/subdomains/supporting/payment/repositories/special-external-account.repository.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { BaseRepository } from 'src/shared/repositories/base.repository';
 import { EntityManager } from 'typeorm';
-import { SpecialExternalAccount } from './special-external-account.entity';
+import { SpecialExternalAccount } from '../entities/special-external-account.entity';
 
 @Injectable()
 export class SpecialExternalAccountRepository extends BaseRepository<SpecialExternalAccount> {
